Adding dev mode
