+++ core/docker/data/conf/configuration.yml
@@ -26,7 +26,7 @@ server:
 </#if>
 
 allocation:
-  strategy: ${env.ALLOCATION_STRATEGY!'CLUSTER'}
+  strategy: ${env.ALLOCATION_STRATEGY!'NONE'}
 
 auth:
   allowedClockSkew: 1 minute
