+++ core/src/main/java/io/paradoxical/cassieq/bundles/GuiceBundleProvider.java
@@ -54,3 +54,4 @@ public class GuiceBundleProvider {
         return builder.build();
     }
 }
+
