+++ core/src/test/java/io/paradoxical/cassieq/unittests/DbTestBase.java
@@ -4,7 +4,7 @@ import com.datastax.driver.core.Session;
 import com.godaddy.logging.Logger;
 import com.google.inject.Injector;
 import io.paradoxical.cassieq.ServiceConfiguration;
-import io.paradoxical.cassieq.unittests.data.CqlDb;
+import io.paradoxical.cassieq.dataAccess.CqlDb;
 import io.paradoxical.cassieq.unittests.modules.InMemorySessionProvider;
 
 import static com.godaddy.logging.LoggerFactory.getLogger;
